Move DownloadStatusPanel to own file
