+++ cmd/launcher/gui/gui.go
@@ -11,19 +11,6 @@ import (
 	"github.com/setlog/trivrost/pkg/misc"
 )
 
-type DownloadStatusPanel struct {
-	*ui.Box
-
-	labelStage *ui.Label
-
-	barTotalProgress *ui.ProgressBar
-	labelStatus      *ui.Label
-
-	progressPrevious, progressCurrent, progressTarget uint64 // Whether these refer to amount of bytes downloaded or something else depends on the current GUI stage.
-	currentProblemMessage                             string
-	stage                                             Stage
-}
-
 var (
 	window              *ui.Window
 	windowTitle         string
@@ -179,7 +166,7 @@ func Main(ctx context.Context, cancelFunc func(), title string, showMainWindow b
 			return false
 		})
 
-		panelDownloadStatus = makeContent()
+		panelDownloadStatus = newDownloadStatusPanel()
 		window.SetChild(panelDownloadStatus)
 		window.SetMargined(true)
 
@@ -199,24 +186,4 @@ func Main(ctx context.Context, cancelFunc func(), title string, showMainWindow b
 
 		guiInitWaitGroup.Done()
 	})
-}
-
-func makeContent() *DownloadStatusPanel {
-	panel := &DownloadStatusPanel{Box: ui.NewVerticalBox()}
-	panel.SetPadded(true)
-
-	panel.labelStage = ui.NewLabel("Initializing...")
-	panel.barTotalProgress = ui.NewProgressBar()
-	panel.barTotalProgress.SetValue(-1)
-	panel.labelStatus = ui.NewLabel("")
-
-	panel.Box.Append(panel.labelStage, false)
-	panel.Box.Append(panel.barTotalProgress, false)
-
-	hBox := ui.NewHorizontalBox()
-	hBox.Append(panel.labelStatus, false)
-	hBox.Append(newLinkLabel("Show logs...", showLogFolder), true)
-	panel.Box.Append(hBox, false)
-
-	return panel
 }
